feat(T-153–T-172): performance, edge-case hardening, compile fixes
T-153: TransactionsController.loadPage() pagination (50 items/page)
T-154/T-155: TransactionHistoryScreen — scroll-based auto-load within 200px of bottom
T-156: _isFiltering state scaffolded in TransactionHistoryScreen
T-157: startup already lazy via ChangeNotifierProvider create: callbacks
T-158: DashboardController — skip all migrations when configVersion == latest
T-159: Splash minimum wait reduced 1200ms → 800ms
T-160: SmsAutoScanService — exact amount + normalised merchant + ±6h window duplicate
T-161: SMS processed hash set (max 1000, evict oldest) to skip re-processed SMS
T-162/T-163: RepaintBoundary already present; verified in investments + dashboard
T-164: const constructors already present in separator/date widgets
T-165/T-166: DB schema v2 — compound indexes idx_tx_acc_dt and idx_tx_type_dt
T-167: TransactionDetailsContent — shows "Deleted account" from metadata
T-168: GoalsScreen — shows "Overdue by Xd" (amber) for negative daysRemaining
T-169: BudgetsController — multi-period rollover loop until current period
T-170: FdWizardScreen — warns when maturity date is in the past
T-171: LoanWizard — validates EMI × tenure ≥ principal before saving
T-172: VoiceFillEngine — _inFlight guard cancels previous call on rapid double-tap

Fix: investments_controller XIRR isolate type cast (int→double for millisecondsSinceEpoch)
Fix: quick_entry_sheet Tooltip import from material.dart
Fix: fd_details_screen const InvestmentTypeSelectionModal → non-const

+++ finance_app/lib/logic/investments_controller.dart
@@ -338,7 +338,9 @@ class InvestmentsController with ChangeNotifier, SafeStorageMixin {
 
     return compute(
       _xirrIsolate,
-      cashflows.map((c) => [c.$1.millisecondsSinceEpoch, c.$2]).toList(),
+      cashflows
+          .map((c) => [c.$1.millisecondsSinceEpoch.toDouble(), c.$2])
+          .toList(),
     );
   }
